Move main import procedures back to type
